refactor: Gamer와 Gamers 클래스 생성 및 패키지 리팩터링
