+++ src/main/java/blackjackgame/domain/blackjack/DealerRandomCardDrawStrategy.java
@@ -1,5 +1,7 @@
 package blackjackgame.domain.blackjack;
 
+import blackjackgame.domain.gamers.CardHolder;
+
 public class DealerRandomCardDrawStrategy extends AbstractRandomCardDrawStrategy {
     private final CardHolder dealer;
