+++ src/main/java/blackjackgame/domain/blackjack/PlayerRandomCardDrawStrategy.java
@@ -1,5 +1,7 @@
 package blackjackgame.domain.blackjack;
 
+import blackjackgame.domain.gamers.CardHolder;
+
 public class PlayerRandomCardDrawStrategy extends AbstractRandomCardDrawStrategy {
     private final CardHolder player;
