+++ src/test/java/blackjackgame/domain/blackjack/GameResultCalculatorTest.java
@@ -14,6 +14,7 @@ import static blackjackgame.domain.card.CardType.DIAMOND;
 import static blackjackgame.domain.card.CardType.HEART;
 
 import blackjackgame.domain.card.Card;
+import blackjackgame.domain.gamers.CardHolder;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
