+++ src/test/java/blackjackgame/domain/blackjack/TestDealerCardDrawStrategy.java
@@ -1,5 +1,7 @@
 package blackjackgame.domain.blackjack;
 
+import blackjackgame.domain.gamers.CardHolder;
+
 public class TestDealerCardDrawStrategy extends TestAbstractFirstCardDrawStrategy {
     private final CardHolder player;
